Add w3d1 07-08, w3d2 07. Fixes in w3 files, mostly dangling pointers

+++ week-03/day-2/01.cpp
@@ -5,8 +5,8 @@ using namespace std;
 
 struct Computer {
   string name;
-  float cpu_speed_hz;
-  float ram_size;
+  long int cpu_speed_hz;
+  long int ram_size;
 };
 
 int main() {
